Verify valid provenance and SBOM attestations
These checks are pretty bare bones at the moment as we don't have much
in these attestations that is consistent across all DOI that we can
check for.

There is a lot of duplication at the moment which I hope can be reduced,
but I'm operating at the edge of my rego knowledge here!