Fix running integration tests locally on macOS
Mounting the ~/.docker dir into the container means that the auth
settings in config.json are used. If the user uses docker desktop this
means it will try to invoke the credential helper to get the creds,
which fails as this doesn't exist in the container. We can just remove
this mount as the image we need to pull is public.

Also disable TUF do we don't try to write to ~/.docker/tuf, which is a
good idea anyway as we only want to test the policy in this repo.